+++ server/src/routes/citizen/vehicles.ts
@@ -3,17 +3,17 @@ import { useAuth } from "../../hooks";
 import { Router, Response } from "express";
 import { processQuery } from "../../lib/database";
 import { v4 as uuidv4 } from "uuid";
-import { generateVinNumber } from "../../lib/functions";
+import { generateString } from "../../lib/functions";
 
 const router: Router = Router();
 
 router.get("/:id", useAuth, async (req: IRequest, res: Response) => {
   const { id } = req.params;
 
-  const vehicles = await processQuery(
-    "SELECT * FROM `registered_cars` WHERE `citizen_id` = ? AND `user_id` = ?",
-    [id, req.user?.id]
-  );
+  const vehicles = await processQuery("SELECT * FROM `registered_cars` WHERE `citizen_id` = ? AND `user_id` = ?", [
+    id,
+    req.user?.id,
+  ]);
 
   return res.json({ vehicles, status: "success" });
 });
@@ -21,10 +21,10 @@ router.get("/:id", useAuth, async (req: IRequest, res: Response) => {
 router.get("/i/:id", useAuth, async (req: IRequest, res: Response) => {
   const { id } = req.params;
 
-  const vehicle = await processQuery(
-    "SELECT * FROM `registered_cars` WHERE `id` = ? AND `user_id` = ?",
-    [id, req.user?.id]
-  );
+  const vehicle = await processQuery("SELECT * FROM `registered_cars` WHERE `id` = ? AND `user_id` = ?", [
+    id,
+    req.user?.id,
+  ]);
 
   return res.json({ vehicle: vehicle[0], status: "success" });
 });
@@ -35,15 +35,9 @@ router.post("/", useAuth, async (req: IRequest, res: Response) => {
   if (plate && status && color && vehicle && citizenId) {
     const parsedPlate = plate.replace(/[oO]/g, "0");
 
-    const citizen = await processQuery(
-      "SELECT `full_name`, `business_id` FROM `citizens` WHERE `id` = ?",
-      [citizenId]
-    );
+    const citizen = await processQuery("SELECT `full_name`, `business_id` FROM `citizens` WHERE `id` = ?", [citizenId]);
 
-    const existingPlate = await processQuery(
-      "SELECT `plate` from `registered_cars` WHERE `plate` = ?",
-      [parsedPlate]
-    );
+    const existingPlate = await processQuery("SELECT `plate` from `registered_cars` WHERE `plate` = ?", [parsedPlate]);
 
     if (existingPlate[0]) {
       return res.json({
@@ -85,22 +79,11 @@ router.post("/", useAuth, async (req: IRequest, res: Response) => {
     }
 
     const id = uuidv4();
-    const vin = generateVinNumber();
+    const vin = generateString(17);
 
     await processQuery(
       "INSERT INTO `registered_cars` (`id`, `owner`, `citizen_id`, `vehicle`, `vin_number`, `in_status`, `plate`, `color`, `user_id`, `business_id`) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?)",
-      [
-        id,
-        citizen[0].full_name,
-        citizenId,
-        vehicle,
-        vin,
-        status,
-        parsedPlate,
-        color,
-        req.user?.id,
-        businessId,
-      ]
+      [id, citizen[0].full_name, citizenId, vehicle, vin, status, parsedPlate, color, req.user?.id, businessId],
     );
 
     return res.json({ status: "success", citizenId });
@@ -114,9 +97,7 @@ router.post("/transfer/:id", useAuth, async (req: IRequest, res: Response) => {
   const { plate, ownerId } = req.body;
 
   if (plate && ownerId) {
-    const citizen = await processQuery("SELECT `full_name` FROM `citizens` WHERE `id` = ?", [
-      ownerId,
-    ]);
+    const citizen = await processQuery("SELECT `full_name` FROM `citizens` WHERE `id` = ?", [ownerId]);
     const vehicle = await processQuery("SELECT `id` FROM `registered_cars` WHERE `id` = ?", [id]);
 
     if (!citizen[0]) {
@@ -127,10 +108,12 @@ router.post("/transfer/:id", useAuth, async (req: IRequest, res: Response) => {
       return res.json({ error: "Vehicle does not exist", status: "error" });
     }
 
-    await processQuery(
-      "UPDATE `registered_cars` SET `plate` = ?, `citizen_id` = ?, `owner` = ? WHERE `id` = ?",
-      [plate, ownerId, citizen[0].full_name, id]
-    );
+    await processQuery("UPDATE `registered_cars` SET `plate` = ?, `citizen_id` = ?, `owner` = ? WHERE `id` = ?", [
+      plate,
+      ownerId,
+      citizen[0].full_name,
+      id,
+    ]);
 
     return res.json({ status: "success" });
   } else {
@@ -143,10 +126,12 @@ router.put("/:id", useAuth, async (req: IRequest, res: Response) => {
   const { color, status } = req.body;
 
   if (color && status) {
-    await processQuery(
-      "UPDATE `registered_cars` SET `color` = ?, `in_status` = ? WHERE `id` = ? AND `user_id` = ?",
-      [color, status, id, req.user?.id]
-    );
+    await processQuery("UPDATE `registered_cars` SET `color` = ?, `in_status` = ? WHERE `id` = ? AND `user_id` = ?", [
+      color,
+      status,
+      id,
+      req.user?.id,
+    ]);
 
     return res.json({ status: "success" });
   } else {
@@ -157,10 +142,7 @@ router.put("/:id", useAuth, async (req: IRequest, res: Response) => {
 router.put("/report-stolen/:vehicleId", useAuth, async (req: IRequest, res: Response) => {
   const { vehicleId } = req.params;
 
-  await processQuery("UPDATE `registered_cars` SET `in_status` = ? WHERE `id` = ?", [
-    "Reported stolen",
-    vehicleId,
-  ]);
+  await processQuery("UPDATE `registered_cars` SET `in_status` = ? WHERE `id` = ?", ["Reported stolen", vehicleId]);
 
   return res.json({
     status: "success",
@@ -172,10 +154,10 @@ router.delete("/:citizenId/:vehicleId", useAuth, async (req: IRequest, res: Resp
 
   await processQuery("DELETE FROM `registered_cars` WHERE `id` = ?", [vehicleId]);
 
-  const vehicles = await processQuery(
-    "SELECT * FROM `registered_cars` WHERE `citizen_id` = ? AND `user_id` = ?",
-    [citizenId, req.user?.id]
-  );
+  const vehicles = await processQuery("SELECT * FROM `registered_cars` WHERE `citizen_id` = ? AND `user_id` = ?", [
+    citizenId,
+    req.user?.id,
+  ]);
 
   return res.json({
     status: "success",