+++ server/src/routes/citizen/weapons.ts
@@ -3,34 +3,28 @@ import { useAuth } from "../../hooks";
 import { Router, Response } from "express";
 import { processQuery } from "../../lib/database";
 import { v4 as uuidv4 } from "uuid";
-import { generateSerialNumber } from "../../lib/functions";
+import { generateString } from "../../lib/functions";
 
 const router: Router = Router();
 
 router.get("/:id", useAuth, async (req: IRequest, res: Response) => {
   const { id } = req.params;
-  const weapons = await processQuery(
-    "SELECT * FROM `registered_weapons` WHERE `citizen_id` = ?",
-    [id]
-  );
+  const weapons = await processQuery("SELECT * FROM `registered_weapons` WHERE `citizen_id` = ?", [id]);
 
   return res.json({ status: "success", weapons });
 });
 
 router.post("/", useAuth, async (req: IRequest, res: Response) => {
   const { weapon, citizenId, status } = req.body;
-  const citizen = await processQuery(
-    "SELECT `full_name` FROM `citizens` WHERE `id` = ?",
-    [citizenId]
-  );
+  const citizen = await processQuery("SELECT `full_name` FROM `citizens` WHERE `id` = ?", [citizenId]);
 
   if (weapon && citizenId && status) {
     const id = uuidv4();
-    const serial = generateSerialNumber();
+    const serial = generateString(10);
 
     await processQuery(
       "INSERT INTO `registered_weapons` (`id`, `owner`, `citizen_id`, `weapon`, `serial_number`, `status`, `user_id`) VALUES (?, ?, ?, ?, ?, ?, ?)",
-      [id, citizen[0].full_name, citizenId, weapon, serial, status, req.user?.id]
+      [id, citizen[0].full_name, citizenId, weapon, serial, status, req.user?.id],
     );
 
     return res.json({ status: "success", citizenId });
@@ -39,24 +33,14 @@ router.post("/", useAuth, async (req: IRequest, res: Response) => {
   }
 });
 
-router.delete(
-  "/:citizenId/:weaponId",
-  useAuth,
-  async (req: IRequest, res: Response) => {
-    const { citizenId, weaponId } = req.params;
+router.delete("/:citizenId/:weaponId", useAuth, async (req: IRequest, res: Response) => {
+  const { citizenId, weaponId } = req.params;
 
-    await processQuery(
-      "DELETE FROM `registered_weapons` WHERE `id` = ? AND `citizen_id` = ?",
-      [weaponId, citizenId]
-    );
+  await processQuery("DELETE FROM `registered_weapons` WHERE `id` = ? AND `citizen_id` = ?", [weaponId, citizenId]);
 
-    const weapons = await processQuery(
-      "SELECT * FROM `registered_weapons` WHERE `citizen_id` = ?",
-      [citizenId]
-    );
+  const weapons = await processQuery("SELECT * FROM `registered_weapons` WHERE `citizen_id` = ?", [citizenId]);
 
-    return res.json({ status: "success", weapons });
-  }
-);
+  return res.json({ status: "success", weapons });
+});
 
 export default router;